Extract edge output to proper module
