+++ freqtrade/optimize/optimize_reports/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa: F401
-from freqtrade.optimize.optimize_reports.bt_output import (show_backtest_result,
+from freqtrade.optimize.optimize_reports.bt_output import (generate_edge_table,
+                                                           show_backtest_result,
                                                            show_backtest_results,
                                                            show_sorted_pairlist,
                                                            text_table_add_metrics,
@@ -11,7 +12,7 @@ from freqtrade.optimize.optimize_reports.bt_storage import (store_backtest_analy
                                                             store_backtest_stats)
 from freqtrade.optimize.optimize_reports.optimize_reports import (
     generate_all_periodic_breakdown_stats, generate_backtest_stats, generate_daily_stats,
-    generate_edge_table, generate_exit_reason_stats, generate_pair_metrics,
-    generate_periodic_breakdown_stats, generate_rejected_signals, generate_strategy_comparison,
-    generate_strategy_stats, generate_tag_metrics, generate_trade_signal_candles,
-    generate_trading_stats, generate_wins_draws_losses)
+    generate_exit_reason_stats, generate_pair_metrics, generate_periodic_breakdown_stats,
+    generate_rejected_signals, generate_strategy_comparison, generate_strategy_stats,
+    generate_tag_metrics, generate_trade_signal_candles, generate_trading_stats,
+    generate_wins_draws_losses)
