+++ freqtrade/optimize/optimize_reports/bt_output.py
@@ -378,3 +378,28 @@ def show_sorted_pairlist(config: Config, backtest_stats: Dict):
                 if result["key"] != 'TOTAL':
                     print(f'"{result["key"]}",  // {result["profit_mean"]:.2%}')
             print("]")
+
+
+def generate_edge_table(results: dict) -> str:
+    floatfmt = ('s', '.10g', '.2f', '.2f', '.2f', '.2f', 'd', 'd', 'd')
+    tabular_data = []
+    headers = ['Pair', 'Stoploss', 'Win Rate', 'Risk Reward Ratio',
+               'Required Risk Reward', 'Expectancy', 'Total Number of Trades',
+               'Average Duration (min)']
+
+    for result in results.items():
+        if result[1].nb_trades > 0:
+            tabular_data.append([
+                result[0],
+                result[1].stoploss,
+                result[1].winrate,
+                result[1].risk_reward_ratio,
+                result[1].required_risk_reward,
+                result[1].expectancy,
+                result[1].nb_trades,
+                round(result[1].avg_trade_duration)
+            ])
+
+    # Ignore type as floatfmt does allow tuples but mypy does not know that
+    return tabulate(tabular_data, headers=headers,
+                    floatfmt=floatfmt, tablefmt="orgtbl", stralign="right")
